before chat rooms

+++ base/views.py
@@ -1,8 +1,12 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -11,6 +15,9 @@ from .forms import RoomForm
 #          {'id':3,'name':'room3','capacity':30}
 #          ]
 def loginpage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method == 'POST':
         username = request.POST.get('username').lower()
         password = request.POST.get('password')
@@ -20,8 +27,36 @@ def loginpage(request):
         except:
             messages.error(request, 'User does not exist')
         #return redirect('home')
-    context = {}
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or Password is incorrect')
+    context = {'page':page}
     return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    #page = 'register'
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            #form.save()
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('login')
+        else:
+            messages.error(request, 'An error occured during registeration')
+    context = {'form':form}
+    return render(request, 'base/login_register.html', context)
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -38,6 +73,7 @@ def room(request,pk):
     room = Room.objects.get(id=pk)
     context = {'room':room}
     return render(request, 'base/room.html',context)
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -48,9 +84,14 @@ def createRoom(request):
             return redirect('home')
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You have no permission for this action')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -58,8 +99,11 @@ def updateRoom(request,pk):
             return redirect('home')
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You have no permission for this action')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
